moved model_configs out of container_registry config
